Interfaces to prepare a "regular" version when spring is not involved
